fix(ios,survival): SurvivalPhraseStaffSnapshotのEquatable合成エラーを修正する
opacityをDoubleにしてFoundationのみのファイルでも型合成できるようにする。

+++ ios/Jazzify/Survival/Phrase/SurvivalPhraseStaffView.swift
@@ -17,7 +17,7 @@ struct SurvivalPhraseStaffView: View {
             singleMeasureLayout: false,
             hideChordLabels: false,
             noteCollisionLayout: .anchorLow,
-            unpressedNoteOpacity: snapshot.unpressedNoteOpacity
+            unpressedNoteOpacity: CGFloat(snapshot.unpressedNoteOpacity)
         )
         .scaleEffect(1.35, anchor: .top)
     }
